fixed: inconsistencies between entity id type and dao id typing (int/Integer)

+++ src/lib/dao/RequestDAO.java
@@ -4,7 +4,7 @@ import lib.entities.Request;
 import lib.interfaces.IRequestDAO;
 
 
-public class RequestDAO extends GenericDaoImpl<Request, Long> implements IRequestDAO {
+public class RequestDAO extends GenericDaoImpl<Request, Integer> implements IRequestDAO {
 	private static final long serialVersionUID = 38783239783134245L;
 
 }
