+++ src/lib/interfaces/IRequestDAO.java
@@ -3,6 +3,6 @@ package lib.interfaces;
 import lib.entities.Request;
 
 
-public interface IRequestDAO extends IGenericDAO<Request, Long> {
+public interface IRequestDAO extends IGenericDAO<Request, Integer> {
 
 }
